display: remove separator on horizontal split
when switching from vertical split to horizontal split we need to remove
the vertical separator. this fixes a bug introduced in 39acb5a.

Signed-off-by: Sven Wegener <sven.wegener@stealer.net>

Add test for toggling of vertical-split. It doesn't fail before this fix
is applied since the visual artifacts of the bug are probably caused by
ncurses lazy screen refreshing.

Tested-by: Vivien Didelot <vivien.didelot@gmail.com>
Signed-off-by: Jonas Fonseca <fonseca@diku.dk>