require_git_version: allow multiple digits in version
The version check code, in require_git_version(), currently asserts that
git v2.10.0 is less than v2.5.0, since it sees the minor version as the
single digit 1, rather than the number 10. This leads to the worktree
tests being incorrectly skipped. In order to fix the version check, the
regular expressions used to isolate the major/minor version numbers are
modified to allow multiple digits.

Signed-off-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Jonas Fonseca <jonas.fonseca@gmail.com>