save-view: fix some printf format specifiers
Commit 175f658781 ("Test diff stat parsing", 10-07-2017) added a new
'save-view' command to dump the view data as a facility to help in
debugging. In order to dump the data, some variables of type size_t
are printed using an '%ld' fprintf() format specifier. On 64-bit
platforms this is not a problem, since size_t is effectively the
same as an 'unsigned long int'. However, on 32-bit systems, this will
cause compiler warnings to be issued, since size_t is effectively an
'unsigned int' type. In order to print a size_t variable on both
platforms, without provoking compiler warnings, use the '%zu' format
specifier.

Signed-off-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Jonas Fonseca <jonas.fonseca@gmail.com>